backend/requirements.txt
fastapi==0.115.9
uvicorn==0.30.6
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.0
bcrypt==4.2.0
python-multipart==0.0.7
groq==0.13.6
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
pymupdf==1.24.12
requests==2.32.3
email-validator==2.2.3
